% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneRanking.R
\name{assignGeneData}
\alias{assignGeneData}
\title{Calculate gene rank}
\usage{
assignGeneData(object, alpha_cutoff)
}
\arguments{
\item{object}{}
}
\value{
object
}
\description{
Calculate gene rank
}
\examples{
load(system.file("data", "gscreend_experiment.RData", package = "gscreend"))
# pse is a object of PoolScreenExp class
assignGeneData(pse, alpha_cutoff=0.05)
}
