% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrnaRanking.R
\name{calculateIntervalFits}
\alias{calculateIntervalFits}
\title{Calculate fit parameters for every subset of data}
\usage{
calculateIntervalFits(object, quant1, quant2)
}
\arguments{
\item{object}{}
}
\value{
object
}
\description{
Calculate fit parameters for every subset of data
}
