% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{calculateLFC}
\alias{calculateLFC}
\title{Calculate log fold changes}
\usage{
calculateLFC(object)
}
\arguments{
\item{object}{PoolScreenExp object}
}
\value{
object
}
\description{
Calculate log fold changes
}
\examples{
pse <- readRDS(system.file("extdata", "gscreend_experiment.RData",
package = "gscreend"))
# pse is a of PoolScreenExp class
norm_pse <- normalizePoolScreenExp(pse)
calculateLFC(norm_pse)
}
