% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInput.R
\name{createPoolScreenExpFromSE}
\alias{createPoolScreenExpFromSE}
\title{Create PoolScreen Experiment from summarized experiment}
\usage{
createPoolScreenExpFromSE(data)
}
\arguments{
\item{data}{}
}
\value{
object
}
\description{
Create PoolScreen Experiment from summarized experiment
}
