% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrnaRanking.R
\name{fit_least_quantile}
\alias{fit_least_quantile}
\title{Fit paramters for skew normal distribution}
\usage{
fit_least_quantile(LFC, quant1, quant2)
}
\arguments{
\item{LFC}{logarithmic fold changes of gRNA counts}

\item{quant1}{lower quantile for least quantile of squares regression
(default: 0.1)}

\item{quant2}{upper quantile for least quantile of squares regression
(default: 0.9)}
}
\value{
fit_skewnorm
}
\description{
Fit paramters for skew normal distribution
}
