% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{GeneData,PoolScreenExp-method}
\alias{GeneData,PoolScreenExp-method}
\title{Accessor function for the Gene slot of the PoolScreenExp class}
\usage{
\S4method{GeneData}{PoolScreenExp}(x)
}
\arguments{
\item{x}{PoolScreenExp object}
}
\value{
Gene slot of the object
}
\description{
Accessor function for the Gene slot of the PoolScreenExp class
}
\examples{
# import a PoolScreenExp object that has been generated using
# RunGscreend()
pse_an <- readRDS(
system.file('extdata', 'gscreend_analysed_experiment.RData',
package = 'gscreend'))

GeneData(pse_an)

}
