% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{GeneData}
\alias{GeneData}
\title{GeneData: set and retrieve GeneData of PoolScreenExp}
\usage{
GeneData(x)
}
\arguments{
\item{x}{PoolScreenExp object}
}
\value{
Gene slot of the object
}
\description{
GeneData: set and retrieve GeneData of PoolScreenExp
}
\examples{
# import a PoolScreenExp object that has been generated using
# RunGscreend()
pse_an <- readRDS(
system.file('extdata', 'gscreend_analysed_experiment.RData',
package = 'gscreend'))

GeneData(pse_an)

}
