% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalizePoolScreenExp}
\alias{normalizePoolScreenExp}
\title{Normalize raw count}
\usage{
normalizePoolScreenExp(object)
}
\arguments{
\item{object}{}
}
\value{
object
}
\description{
Normalize raw count
}
\examples{
load(system.file("data", "gscreend_experiment.RData", package = "gscreend"))
# pse is a of PoolScreenExp class
normalizePoolScreenExp(pse)
}
