% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotModelParameters}
\alias{plotModelParameters}
\title{Plot model parameters from the fitting}
\usage{
plotModelParameters(object)
}
\arguments{
\item{object}{PoolScreenExp object}
}
\value{
plot
}
\description{
Plot model parameters from the fitting
}
\examples{
pse <- readRDS(system.file("extdata", "gscreend_experiment.RData",
package = "gscreend"))
# pse is a poolscreen Experiment with two replicates R1 and R2
pse_an <- RunGscreend(pse)
plotModelParameters(pse_an)
}
