% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotReplicateCorrelation}
\alias{plotReplicateCorrelation}
\title{Plot replicate correlation}
\usage{
plotReplicateCorrelation(object, rep1 = "R1", rep2 = "R2")
}
\arguments{
\item{object}{}

\item{rep1}{}

\item{rep2}{}
}
\value{
plot
}
\description{
Plot replicate correlation
}
