% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{ResultsTable}
\alias{ResultsTable}
\title{Extract a results table}
\usage{
ResultsTable(object, direction = "negative")
}
\arguments{
\item{object}{PoolScreenExp object}

\item{direction}{Whether the table should contain information on positive or
negative fold changes ["negative"| "positive"]}
}
\value{
plot
}
\description{
Extract a results table
}
\examples{
# import a PoolScreenExp object that has been generated using
# RunGscreend()
pse_an <- readRDS(
system.file("extdata", "gscreend_analysed_experiment.RData",
package = "gscreend"))
ResultsTable(pse_an, direction = "negative")

}
