% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sgRNAData}
\alias{sgRNAData}
\title{sgRNAData: set and retrieve sgRNAData of PoolScreenExp}
\usage{
sgRNAData(x)
}
\arguments{
\item{x}{PoolScreenExp object}
}
\value{
sgRNA slot of the object
}
\description{
sgRNAData: set and retrieve sgRNAData of PoolScreenExp
}
\examples{
# import a PoolScreenExp object that has been generated using
# RunGscreend()
pse_an <- readRDS(
system.file('extdata', 'gscreend_analysed_experiment.RData',
package = 'gscreend'))

sgRNAData(pse_an)

}
