% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneRanking.R
\name{assignGeneData}
\alias{assignGeneData}
\title{Calculate gene rank}
\usage{
assignGeneData(object, alpha_cutoff, n_cores)
}
\arguments{
\item{object}{PoolScreenExp object}

\item{alpha_cutoff}{alpha cutoff for alpha-RRA (default: 0.05)}

\item{n_cores}{number of cores to be used (default: 1)}
}
\value{
object
}
\description{
Calculate gene rank
}
