% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrnaRanking.R
\name{calculateIntervalFits}
\alias{calculateIntervalFits}
\title{Calculate fit parameters for every subset of data}
\usage{
calculateIntervalFits(object, quant1, quant2)
}
\arguments{
\item{object}{PoolScreenExp object}

\item{quant1}{lower quantile for least quantile of squares regression
(default: 0.1)}

\item{quant2}{upper quantile for least quantile of squares regression
(default: 0.9)}
}
\value{
object
}
\description{
Calculate fit parameters for every subset of data
}
\keyword{internal}
