% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInput.R
\name{createPoolScreenExpFromSE}
\alias{createPoolScreenExpFromSE}
\title{Create PoolScreenExp Experiment from summarized experiment}
\usage{
createPoolScreenExpFromSE(data)
}
\arguments{
\item{data}{SummarizedExperiment object containing gRNA level data}
}
\value{
object
}
\description{
Create PoolScreenExp Experiment from summarized experiment
}
\keyword{internal}
