% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrnaRanking.R
\name{defineFittingIntervals}
\alias{defineFittingIntervals}
\title{Calculate intervall limits for splitting data into subsets}
\usage{
defineFittingIntervals(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Calculate intervall limits for splitting data into subsets
}
