% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrnaRanking.R
\name{fit_least_quantile}
\alias{fit_least_quantile}
\title{Fit paramters for skew normal distribution}
\usage{
fit_least_quantile(LFC)
}
\arguments{
\item{LFC}{}
}
\value{
fit_skewnorm
}
\description{
Fit paramters for skew normal distribution
}
