% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalizePoolScreenExp}
\alias{normalizePoolScreenExp}
\title{Normalize raw count}
\usage{
normalizePoolScreenExp(object)
}
\arguments{
\item{object}{PoolScreenExp object}
}
\value{
object
}
\description{
Normalize raw count
}
\examples{
pse <- readRDS(system.file("extdata", "gscreend_experiment.RData",
package = "gscreend"))
# pse is a PoolScreenExp
normalizePoolScreenExp(pse)
}
