% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotModelParameters}
\alias{plotModelParameters}
\title{Plot model parameters from the fitting}
\usage{
plotModelParameters(object)
}
\arguments{
\item{object}{}
}
\value{
plot
}
\description{
Plot model parameters from the fitting
}
\examples{
load(system.file("data", "gscreend_experiment.RData", package = "gscreend"))
# pse is a object of PoolScreenExp class
plotModelParameters(pse)
}
