% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotModelParameters}
\alias{plotModelParameters}
\title{Plot model parameters from the fitting}
\usage{
plotModelParameters(object)
}
\arguments{
\item{object}{PoolScreenExp object}
}
\value{
plot
}
\description{
Plot model parameters from the fitting
}
\examples{
# import a PoolScreenExp object that has been generated using
# RunGscreend()
pse_an <- readRDS(
system.file("extdata", "gscreend_analysed_experiment.RData",
package = "gscreend"))
plotModelParameters(pse_an)
}
