% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotReplicateCorrelation}
\alias{plotReplicateCorrelation}
\title{Plot replicate correlation}
\usage{
plotReplicateCorrelation(object, rep1 = "R1", rep2 = "R2")
}
\arguments{
\item{rep2}{}
}
\value{
plot
}
\description{
Plot replicate correlation
}
\examples{
load(system.file("data", "gscreend_experiment.RData", package = "gscreend"))
# pse is a poolscreen Experiment with two replicates R1 and R2
plotReplicateCorrelation(pse, rep1 = "R1", rep2 = "rep2")
}
