% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotReplicateCorrelation}
\alias{plotReplicateCorrelation}
\title{Plot replicate correlation}
\usage{
plotReplicateCorrelation(object, rep1 = "R1", rep2 = "R2")
}
\arguments{
\item{object}{PoolScreenExp object}

\item{rep1}{Name of replicate 1}

\item{rep2}{Name of replicate 2}
}
\value{
replicate_plot
}
\description{
Plot replicate correlation
}
\examples{
# import a PoolScreenExp object that has been generated using RunGscreend()
pse_an <- readRDS(
system.file("extdata", "gscreend_analysed_experiment.RData",
package = "gscreend"))
plotReplicateCorrelation(pse_an, rep1 = "R1", rep2 = "R2")

}
