% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotReplicateCorrelation}
\alias{plotReplicateCorrelation}
\title{Plot replicate correlation}
\usage{
plotReplicateCorrelation(object, rep1 = "R1", rep2 = "R2")
}
\arguments{
\item{object}{PoolScreenExp object}

\item{rep1}{Name of replicate 1}

\item{rep2}{Name of replicate 2}
}
\value{
replicate_plot
}
\description{
Plot replicate correlation
}
\examples{
pse <- readRDS(system.file("extdata", "gscreend_experiment.RData",
package = "gscreend"))
# pse is a poolscreen Experiment with two replicates R1 and R2
pse_an <- RunGscreend(pse)
plotReplicateCorrelation(pse_an, rep1 = "R1", rep2 = "R2")

}
